requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
langchain-groq
chromadb
pandas